fix eval integration
