+++ PaddleCV/PaddleDetection/ppdet/utils/coco_eval.py
@@ -20,13 +20,12 @@ from __future__ import unicode_literals
 import os
 import sys
 import json
-import cv2
-import numpy as np
 import matplotlib
 matplotlib.use('Agg')
 from pycocotools.coco import COCO
 from pycocotools.cocoeval import COCOeval
-import pycocotools.mask as mask_util
+
+from .eval_helper import bbox2out, mask2out, proposal2out
 
 import logging
 logger = logging.getLogger(__name__)
@@ -34,20 +33,10 @@ logger = logging.getLogger(__name__)
 __all__ = [
     'bbox_eval',
     'mask_eval',
-    'bbox2out',
-    'mask2out',
     'get_category_info',
     'proposal_eval',
     'cocoapi_eval',
 ]
-
-
-def clip_bbox(bbox):
-    xmin = max(min(bbox[0], 1.), 0.)
-    ymin = max(min(bbox[1], 1.), 0.)
-    xmax = max(min(bbox[2], 1.), 0.)
-    ymax = max(min(bbox[3], 1.), 0.)
-    return xmin, ymin, xmax, ymax
 
 
 def proposal_eval(results, anno_file, outfile, max_dets=(100, 300, 1000)):
@@ -148,195 +137,6 @@ def cocoapi_eval(jsonfile,
     coco_eval.accumulate()
     coco_eval.summarize()
     return coco_eval.stats
-
-
-def proposal2out(results, is_bbox_normalized=False):
-    xywh_res = []
-    for t in results:
-        bboxes = t['proposal'][0]
-        lengths = t['proposal'][1][0]
-        im_ids = np.array(t['im_id'][0])
-        if bboxes.shape == (1, 1) or bboxes is None:
-            continue
-
-        k = 0
-        for i in range(len(lengths)):
-            num = lengths[i]
-            im_id = int(im_ids[i][0])
-            for j in range(num):
-                dt = bboxes[k]
-                xmin, ymin, xmax, ymax = dt.tolist()
-
-                if is_bbox_normalized:
-                    xmin, ymin, xmax, ymax = \
-                            clip_bbox([xmin, ymin, xmax, ymax])
-                    w = xmax - xmin
-                    h = ymax - ymin
-                else:
-                    w = xmax - xmin + 1
-                    h = ymax - ymin + 1
-
-                bbox = [xmin, ymin, w, h]
-                coco_res = {
-                    'image_id': im_id,
-                    'category_id': 1,
-                    'bbox': bbox,
-                    'score': 1.0
-                }
-                xywh_res.append(coco_res)
-                k += 1
-    return xywh_res
-
-
-def bbox2out(results, clsid2catid, is_bbox_normalized=False):
-    """
-    Args:
-        results: request a dict, should include: `bbox`, `im_id`,
-                 if is_bbox_normalized=True, also need `im_shape`.
-        clsid2catid: class id to category id map of COCO2017 dataset.
-        is_bbox_normalized: whether or not bbox is normalized.
-    """
-    xywh_res = []
-    for t in results:
-        bboxes = t['bbox'][0]
-        lengths = t['bbox'][1][0]
-        im_ids = np.array(t['im_id'][0])
-        if bboxes.shape == (1, 1) or bboxes is None:
-            continue
-
-        k = 0
-        for i in range(len(lengths)):
-            num = lengths[i]
-            im_id = int(im_ids[i][0])
-            for j in range(num):
-                dt = bboxes[k]
-                clsid, score, xmin, ymin, xmax, ymax = dt.tolist()
-                catid = clsid2catid[clsid]
-
-                if is_bbox_normalized:
-                    xmin, ymin, xmax, ymax = \
-                            clip_bbox([xmin, ymin, xmax, ymax])
-                    w = xmax - xmin
-                    h = ymax - ymin
-                    im_height, im_width = t['im_shape'][0][i].tolist()
-                    xmin *= im_width
-                    ymin *= im_height
-                    w *= im_width
-                    h *= im_height
-                else:
-                    w = xmax - xmin + 1
-                    h = ymax - ymin + 1
-
-                bbox = [xmin, ymin, w, h]
-                coco_res = {
-                    'image_id': im_id,
-                    'category_id': catid,
-                    'bbox': bbox,
-                    'score': score
-                }
-                xywh_res.append(coco_res)
-                k += 1
-    return xywh_res
-
-
-def mask2out(results, clsid2catid, resolution, thresh_binarize=0.5):
-    scale = (resolution + 2.0) / resolution
-
-    segm_res = []
-
-    # for each batch
-    for t in results:
-        bboxes = t['bbox'][0]
-
-        lengths = t['bbox'][1][0]
-        im_ids = np.array(t['im_id'][0])
-        if bboxes.shape == (1, 1) or bboxes is None:
-            continue
-        if len(bboxes.tolist()) == 0:
-            continue
-
-        masks = t['mask'][0]
-
-        s = 0
-        # for each sample
-        for i in range(len(lengths)):
-            num = lengths[i]
-            im_id = int(im_ids[i][0])
-            im_shape = t['im_shape'][0][i]
-
-            bbox = bboxes[s:s + num][:, 2:]
-            clsid_scores = bboxes[s:s + num][:, 0:2]
-            mask = masks[s:s + num]
-            s += num
-
-            im_h = int(im_shape[0])
-            im_w = int(im_shape[1])
-
-            expand_bbox = expand_boxes(bbox, scale)
-            expand_bbox = expand_bbox.astype(np.int32)
-
-            padded_mask = np.zeros(
-                (resolution + 2, resolution + 2), dtype=np.float32)
-
-            for j in range(num):
-                xmin, ymin, xmax, ymax = expand_bbox[j].tolist()
-                clsid, score = clsid_scores[j].tolist()
-                clsid = int(clsid)
-                padded_mask[1:-1, 1:-1] = mask[j, clsid, :, :]
-
-                catid = clsid2catid[clsid]
-
-                w = xmax - xmin + 1
-                h = ymax - ymin + 1
-                w = np.maximum(w, 1)
-                h = np.maximum(h, 1)
-
-                resized_mask = cv2.resize(padded_mask, (w, h))
-                resized_mask = np.array(
-                    resized_mask > thresh_binarize, dtype=np.uint8)
-                im_mask = np.zeros((im_h, im_w), dtype=np.uint8)
-
-                x0 = min(max(xmin, 0), im_w)
-                x1 = min(max(xmax + 1, 0), im_w)
-                y0 = min(max(ymin, 0), im_h)
-                y1 = min(max(ymax + 1, 0), im_h)
-
-                im_mask[y0:y1, x0:x1] = resized_mask[(y0 - ymin):(y1 - ymin), (
-                    x0 - xmin):(x1 - xmin)]
-                segm = mask_util.encode(
-                    np.array(
-                        im_mask[:, :, np.newaxis], order='F'))[0]
-                catid = clsid2catid[clsid]
-                segm['counts'] = segm['counts'].decode('utf8')
-                coco_res = {
-                    'image_id': im_id,
-                    'category_id': catid,
-                    'segmentation': segm,
-                    'score': score
-                }
-                segm_res.append(coco_res)
-    return segm_res
-
-
-def expand_boxes(boxes, scale):
-    """
-    Expand an array of boxes by a given scale.
-    """
-    w_half = (boxes[:, 2] - boxes[:, 0]) * .5
-    h_half = (boxes[:, 3] - boxes[:, 1]) * .5
-    x_c = (boxes[:, 2] + boxes[:, 0]) * .5
-    y_c = (boxes[:, 3] + boxes[:, 1]) * .5
-
-    w_half *= scale
-    h_half *= scale
-
-    boxes_exp = np.zeros(boxes.shape)
-    boxes_exp[:, 0] = x_c - w_half
-    boxes_exp[:, 2] = x_c + w_half
-    boxes_exp[:, 1] = y_c - h_half
-    boxes_exp[:, 3] = y_c + h_half
-
-    return boxes_exp
 
 
 def get_category_info(anno_file=None,
